third_party/WebKit/LayoutTests/inspector/console/console-log-in-xhtml-expected.txt

CONSOLE MESSAGE: line 5: 239
Tests that console message from inline script in xhtml document contains correct script position information.

239:5:16
